added a confirmation to /remove_all_ranks

+++ commands/ranking/remove_all_ranks.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder, EmbedBuilder, PermissionsBitField } = require('discord.js');
+const { SlashCommandBuilder, EmbedBuilder, PermissionsBitField, Colors } = require('discord.js');
 const db = require("../../dbObjects.js")
 const testers = require("../../tester_servers.json");
 
@@ -13,8 +13,35 @@ module.exports = {
     async execute(interaction) {
         await interaction.deferReply()
 
-        const removeCount = await db.Ranks.destroy({where: { guild_id: interaction.guild.id}})
-        const embededReply = new EmbedBuilder().setColor([255, 255, 0]).setDescription(`successfuly removed all ${removeCount} ranks`)
+        const deleteButton = new ButtonBuilder()
+            .setCustomId('delete_all_ranks')
+            .setLabel('Delete')
+            .setStyle(ButtonStyle.Danger)
+
+        const cancel = new ButtonBuilder()
+            .setCustomId('cancel_deleting_all_ranks')
+            .setLabel('Cancel')
+            .setStyle(ButtonStyle.Secondary)
+        const row = new ActionRowBuilder().addComponents(deleteButton, cancel)
+
+        const response = await interaction.editReply({embeds: [new EmbedBuilder().setColor(Colors.Red).setDescription(`Are you sure you want to **delete all** the Rank links?`)], components: [row]})
+
+        const collectorFilter = i => i.customId === 'delete_all_ranks' && i.user.id === interaction.user.id
+        try {
+            const confirmation = await response.awaitMessageComponent({ Filter: collectorFilter, time: 60_000 })
+
+            if (confirmation.customId === 'delete_rank') {
+                const removeCount = await db.Ranks.destroy({where: { guild_id: interaction.guild.id}})
+                return interaction.editReply({embeds: [new EmbedBuilder().setColor(Colors.Red).setDescription(`successfuly removed all ${removeCount} ranks`)], components: []})
+            
+            } else if (confirmation.customId === 'cancel_deleting_all_ranks') {
+                return interaction.editReply({embeds: [new EmbedBuilder().setColor(Colors.Green).setDescription(`The deletion has been cancelled!`)], components: []})
+            }
+        } catch (error) {
+            return interaction.editReply({embeds: [embeded_error.setDescription("No responce was given in within 60 secounds, cancelling!")], components: []})
+        }
+
+        
         interaction.editReply({embeds: [embededReply]})
     }
 }
